# docs/swagger.yaml
basePath: /
definitions:
  requests.CreateTask:
    properties:
      text:
        type: string
    type: object
  responses.Task:
    properties:
      created_at:
        type: string
      duration:
        type: string
      status:
        type: string
      task_id:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Service for processing tasks
  title: Task service
  version: 0.0.1
paths:
  /task:
    post:
      consumes:
      - application/json
      description: Создание задачи с помощь передачи текста задания
      parameters:
      - description: структура запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.CreateTask'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Task'
        "400":
          description: некорректный формат запроса
          schema:
            type: string
        "500":
          description: внутренняя ошибка сервера
          schema:
            type: string
      summary: Создание новой задачи
  /task/{task_id}:
    delete:
      consumes:
      - application/json
      description: Удаление задачи по ее id
      parameters:
      - description: path format
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: некорректный формат запроса
          schema:
            type: string
        "404":
          description: задача не найдена
          schema:
            type: string
        "500":
          description: внутренняя ошибка сервера
          schema:
            type: string
      summary: Удаление задачи
    get:
      consumes:
      - application/json
      description: запрос на получение задачи с помощью ее ID
      parameters:
      - description: path format
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Task'
        "404":
          description: задача не найдена
          schema:
            type: string
        "500":
          description: внутренняя ошибка сервера
          schema:
            type: string
      summary: запрос на получение задачи
  /tasks:
    get:
      consumes:
      - application/json
      description: Возвращает список задач с поддержкой пагинации
      parameters:
      - default: 10
        description: Количество задач на странице
        in: query
        name: limit
        type: integer
      - default: 0
        description: Смещение
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.Task'
            type: array
        "400":
          description: некорректный формат запроса
          schema:
            type: string
        "500":
          description: внутренняя ошибка сервера
          schema:
            type: string
      summary: Получение списка задач
swagger: "2.0"